Added Multi-screen support
FloatingClock can now detect multiple screens and be set to display on
any one of them. Actual device names are not shown as that would require
JNI and over-complicate a simple program, so that's not for me to do.
Program will still crash (giving the appearance of failure to launch) if
the .config file is corrupt / empty / missing options.